requirements.txt
pytest
selenium
webdriver_manager
python-dotenv
pytest-html